requirements.txt
# ML Injury Prediction - Football
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
matplotlib>=3.7.0
seaborn>=0.12.0
kaggle>=1.5.0
fastapi
uvicorn[standard]
google-genai
